Implement the index_range function
This commit adds the function index_range that takes two integer arguments, page and page_size. The function returns a tuple of size two containing a start index and an end index corresponding to the range of indexes to return in a list for those particular pagination parameters.

Fixes #none