Implement FIFO caching system
This commit implements a caching system that uses the First In First Out (FIFO) algorithm.
The class FIFOCache inherits from BaseCaching, and it follows the pycodestyle and PEP8 guidelines.
The put method discards the oldest item in the cache when the number of items in the cache exceeds the MAX_ITEMS attribute
The get method remains the same as the parent calss but with the proper documentation.

Fixes #none